added collision explosion

+++ core/src/com/mygdx/gameengine/interfaces/iMovable.java
@@ -1,7 +1,7 @@
 package com.mygdx.gameengine.interfaces;
 
 public interface iMovable {
-    int getSpeedMultiplier();
-    void setSpeedMultiplier(int speedMultiplier);
+    float getSpeedMultiplier();
+    void setSpeedMultiplier(float speedMultiplier);
     //	public void move(float deltaTime);
 }
